NSXv: Subnet NAT rule configuration
While configuration flag
bind_floatingip_to_all_interfaces = True

the subnet NAT rule should not apply only to external interface, as it
also serves traffic between instances on the same router.
So if instances A and B are connected via the same router, and instance
A is accessing instance B's FIP, traffic should reach instance B with
the router's NAT IP - unless there's a FIP to instance A as well.

Change-Id: Ib312289bed86f8539f593da4a01f800b65f72ac5