NSXP: Forbid multiple listeners per LB pool
The NSX3 driver for LBaaSv2 & Octavia does not support multiple
listeners for the same pool.
This code handles create/update of pool with multiple listeners, and also
create/update of listener with a default pool that is already in use.
For that there is also a need to update the pool DB mapping also when
added/removed to a listener.
Additional related code for session persistence support will be added in
a follow up patch.

Change-Id: Iff4d26f478f5eeef1957a576c98014c6dd0ba635
(cherry picked from commit 4e9a95abfb2b540c098b47059eacc00af8b524d0)