NSX|P+V3: Improve get_ports performance with designate
When the dns extension driver is used, get_ports is very slow
due to a call for plugin.get_netowrk on each port in order to get
the networks dns_domain and avialability zone.
This code uses a direct DB call to get the network dns_domain.
And only if relevant, also uses a DB call to get the availability zone.

Change-Id: I4b0dc4d5c891e82e71dccb5ee2a6d911e8d03e49